requirements.txt
pandas>=1.3.0
openpyxl>=3.0.7
pillow>=8.2.0
PyQt5>=5.15.4
pytest>=6.2.5
pyinstaller>=4.5.1
typing-extensions>=3.10.0
colorlog>=6.0.0
pyyaml>=6.0
psutil>=5.9.0
pywin32>=227
numpy>=1.20.0
setuptools>=58.0.0
wheel>=0.37.0
xlwings>=0.30.0 